LayoutTests/http/wpt/wk-web-animations/timing-model/animation-creation-basic.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Constructing Animation instances with various arguments</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

internals.pauseTimeline(document.timeline);

test(t => {
    assert_own_property(window, "Animation");
}, "The Animation interface is defined.");

test(t => {
    const animation = new Animation;
    assert_equals(animation.effect, null);
    assert_equals(animation.timeline, document.timeline);
}, "Creating Animation without arguments should have a null effect and the document timeline.");

test(t => {
    const animation = new Animation(undefined);
    assert_equals(animation.effect, null);
    assert_equals(animation.timeline, document.timeline);
}, "Creating Animation with an undefined argument should have a null effect and the document timeline.");

test(t => {
    const animation = new Animation(null);
    assert_equals(animation.effect, null);
    assert_equals(animation.timeline, document.timeline);
}, "Creating Animation with a null argument should have a null effect and the document timeline.");

test(t => {
    const effect = new KeyframeEffect(document.body, null);
    const animation = new Animation(effect);
    assert_equals(animation.effect, effect);
    assert_equals(animation.timeline, document.timeline);
}, "Creating Animation with an effect should only set the effect property and have the document timeline.");

</script>
</body>
